New reader features
preReading() now checks if the seperator you gave into the function is valid. The seperator is deemed invalid if it is a letter or a number and preReading() returns an error code 3. Also size calculation was extracted into its another function for cleaner code. Next addition will include : 1- Counting how many lines the csv file has 2- Validating and registering the header line